Remove hardcoded compile-time paths (#51)
